refactor(economy): phase 3 slice 25 cluster merges and extractor scenario migration (#3939)
Merge deal-matcher FRR, validator treasury, and FRR D5 credits modules into
their parent scenario clusters; migrate the province-missing logger test to
a table-driven row with expectLogMessageContains.

Refs #3939
