+++ packages/colonizethis_economy/test/resource_extractor_test.dart
@@ -1,9 +1,6 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_economy/colonizethis_economy.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_test/game_test_fixtures.dart';
 import 'package:colonizethis_test/test.dart';
-import 'package:logger/logger.dart';
 
 import 'package:colonizethis_economy_test_support/colonizethis_economy_test_support.dart';
 
@@ -64,55 +61,10 @@ void main() {
     }
 
     test(
-      'skips connected tile and logs when province missing from region (world-model)',
-      () {
-        final captured = <LogEvent>[];
-        void listener(LogEvent e) => captured.add(e);
-        Logger.addLogListener(listener);
-        addTearDown(() {
-          Logger.removeLogListener(listener);
-          captured.clear();
-        });
-        Logger.level = Level.error;
-        addTearDown(() => Logger.level = Level.off);
-
-        final tileMap = _grainTileMap;
-        final tileState = tileStateFromSpecs(const [
-          TileImprovementSpec('oldWorld|p1|0|0', improvement: 2, roadLevel: 2),
-        ]);
-        final player = Player(
-          id: 'pl1',
-          displayName: 'Spain',
-          isHuman: true,
-          capitalProvinceId: 'oldWorld|p1',
-          capitalTile: const CapitalTile(
-            regionId: 'oldWorld',
-            provinceId: 'oldWorld|p1',
-            x: 0,
-            y: 0,
-          ),
-        );
-        final game = TestFixtures.minimalGame(
-          id: 'g1',
-          capitalTileGrainBonusPerTurn: 0,
-          oldWorld: const RegionData(provinces: []),
-          tileState: tileState,
-          players: [player],
-        );
-        final result = computeExtraction(
-          game: game,
-          tileMapByRegion: {'oldWorld': tileMap},
-          connectivityResult: connectivityFor({'oldWorld|p1|0|0'}),
-          techCapForPlayer: (_) => 4,
-        );
-        expect(result['pl1']!.land['grain'], isNull);
-        expect(
-          captured.any(
-            (e) => e.message.contains('extraction province missing'),
-          ),
-          isTrue,
-        );
-      },
+      provinceMissingFromRegionScenario(grainTileMap: _grainTileMap).label,
+      () => runResourceExtractorScenario(
+        provinceMissingFromRegionScenario(grainTileMap: _grainTileMap),
+      ),
     );
 
     test(
