+++ packages/colonizethis_economy_test_support/lib/src/frr_scenarios/frr_scenarios.dart
@@ -5,4 +5,3 @@ export 'first_right_profit_scenarios.dart';
 export 'frr_credits_expectations.dart';
 export 'frr_credits_scenarios.dart';
 export 'frr_credits_test_support.dart';
-export 'frr_issue_ac_d5_scenarios.dart';
